refactor(ui): drop dead style state module
