+++ apps/ui/src/app/view_style_types.ts
@@ -1,0 +1,3 @@
+export type VisualVariant = "bad" | "muted" | "ok" | "warn";
+
+export type ChoiceCardState = "active" | "draft" | "error";
